refactor(redis): extract legacy scan patterns to redisLegacyPatterns.ts
Move LEGACY_REDIS_SCAN_PATTERNS out of the canonical redisKeys registry
into a dedicated module for migration CLI use. Re-export from redisKeys
for backward compatibility; scripts import directly from redisLegacyPatterns.
